+++ example/example-app/html-preview.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import { get } from 'lodash'
 
 // Material-UI
 import Typography from '@material-ui/core/Typography'
@@ -12,10 +11,6 @@ export default ({editorState=[]}) => (
     {editorState.map(block => {
       if (block.type === 'richtext') {
         return block.data.value
-      } else if (block.type === 'image') {
-        return get(block, 'data.src')
-          ? `<img src="${get(block, 'data.src')}" class="img-size-${get(block, 'data.size')} img-alignment-${get(block, 'data.alignment')}" title="${get(block, 'data.caption')}"/>`
-          : null
       }
       return null
     })}