+++ src/components/editor-block/index.stories.js
@@ -7,17 +7,15 @@ import withReadme from 'storybook-readme/with-readme' //eslint-disable-line
 import { clone } from 'lodash'
 
 // Component imports
-import { EditorBlock, EditorImage } from '../..'
+import { EditorBlock, EditorQuill } from '../..'
 import componentReadme from './README.md'
 
 // Example Config
 const exampleBlock = {
   id: 5,
+  type: 'richtext',
   data: {
-    alignment: 'left',
-    caption: 'Hello World.',
-    size: 'medium',
-    src: 'https://picsum.photos/1000/500'
+    value: ''
   },
   meta: {
     title: 'Input Box'
@@ -48,7 +46,7 @@ class Wrapper extends React.Component {
         block={block}
         onAction={action('onAction')}
       >
-        <EditorImage
+        <EditorQuill
           block={block}
           onChange={this.onChange}
         />