+++ CHANGES.md
@@ -55,6 +55,10 @@ const blockConfig = {
 }
 ```
 
+### Removed
+
+- ExampleApp does not contain example for EditorImage anymore
+
 ## 2018/05/23 [0.4.2][7]
 
 ### Changed